refactor(container): added container/icon files in component
