+++ src/app/page.tsx
@@ -1,12 +1,12 @@
 import React from "react";
-import FooterLinks from "@/containers/auth/footer";
-import AuthXLogo from "@/containers/auth/authXLogo";
-import AuthOptions from "@/containers/auth/authOptions";
+import FooterLinks from "@/components/container/footer";
+import AuthXLogo from "@/components/container/authXLogo";
+import AuthOptions from "@/components/container/authOptions";
 
 const Login = () => {
   return (
-    <div className="h-full w-full flex flex-col">
-      <div className="flex-grow flex lg:flex-row text-[#E7E9EA] bg-black">
+    <div className="h-full w-full flex flex-col text-[#E7E9EA]">
+      <div className="flex-grow flex lg:flex-row bg-black">
         <AuthXLogo />
         <AuthOptions />
       </div>
